fix: - fix dependencies from domain module (dsl, explanation, framework modules) updating gradle.kts files; - remove duplicated folders; - fix old imports; - remove useless folder (explanation old version);
