+++ dsl/src/test/kotlin/dsl/EffectsDSLTest.kt
@@ -2,9 +2,9 @@ package dsl
 import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.matchers.collections.shouldBeIn
 import io.kotest.matchers.shouldBe
-import resources.domain.BlockWorldDomain.Domains.blockWorld
-import resources.domain.BlockWorldDomain.Predicates
-import resources.domain.BlockWorldDomain.predicates
+import domain.BlockWorldDomain.Domains.blockWorld
+import domain.BlockWorldDomain.Predicates
+import domain.BlockWorldDomain.predicates
 
 class EffectsDSLTest : AnnotationSpec() {
     @Test
