+++ dsl/src/test/kotlin/dsl/GoalDSLTest.kt
@@ -2,8 +2,8 @@ package dsl
 
 import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.matchers.shouldBe
-import resources.domain.BlockWorldDomain.Fluents
-import resources.domain.BlockWorldDomain.ProblemsDSL.problemOnAB
+import domain.BlockWorldDomain.Fluents
+import domain.BlockWorldDomain.ProblemsDSL.problemOnAB
 
 class GoalDSLTest : AnnotationSpec() {
     @Test
