+++ dsl/src/test/kotlin/dsl/StatesDSLTest.kt
@@ -2,8 +2,8 @@ package dsl
 
 import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.matchers.shouldBe
-import resources.domain.BlockWorldDomain.ProblemsDSL.problemOnAB
-import resources.domain.BlockWorldDomain.States
+import domain.BlockWorldDomain.ProblemsDSL.problemOnAB
+import domain.BlockWorldDomain.States
 
 class StatesDSLTest : AnnotationSpec() {
     @Test
