+++ dsl/src/test/kotlin/dsl/TypeDSLTest.kt
@@ -3,8 +3,8 @@ package dsl
 import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.matchers.collections.shouldBeIn
 import io.kotest.matchers.shouldBe
-import resources.domain.BlockWorldDomain.Domains.blockWorld
-import resources.domain.BlockWorldDomain.Types
+import domain.BlockWorldDomain.Domains.blockWorld
+import domain.BlockWorldDomain.Types
 
 class TypeDSLTest : AnnotationSpec() {
     @Test
