+++ dsl/src/test/kotlin/dsl/TypesProviderTest.kt
@@ -1,14 +1,15 @@
 package dsl
 
+
 import dsl.provider.TypeProvider
 import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.matchers.collections.shouldBeIn
 import io.kotest.matchers.collections.shouldNotBeIn
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
-import resources.domain.BlockWorldDomain.Domains.blockWorld
-import resources.domain.BlockWorldDomain.Types
-import resources.domain.BlockWorldDomain.types
+import domain.BlockWorldDomain.Domains.blockWorld
+import domain.BlockWorldDomain.Types
+import domain.BlockWorldDomain.types
 
 class TypesProviderTest : AnnotationSpec() {
     val typeProvider = TypeProvider.of(setOf(Types.strings, Types.anything))
