+++ explanation/src/test/kotlin/explanation/Question2Test.kt
@@ -4,10 +4,10 @@ import Plan
 import explanation.impl.Question2
 import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.matchers.shouldBe
-import resources.domain.BlockWorldDomain.Operators.pickA
-import resources.domain.BlockWorldDomain.Operators.pickB
-import resources.domain.BlockWorldDomain.Planners.stripsPlanner
-import resources.domain.BlockWorldDomain.Problems
+import domain.BlockWorldDomain.Operators.pickA
+import domain.BlockWorldDomain.Operators.pickB
+import domain.BlockWorldDomain.Planners.stripsPlanner
+import domain.BlockWorldDomain.Problems
 class Question2Test : AnnotationSpec() {
     @Test
     fun `Remove pickA from the plan to solve the armNotEmpty problem`() {
