+++ framework/build.gradle.kts
@@ -17,4 +17,5 @@ dependencies {
     testImplementation("io.kotest:kotest-runner-junit5:5.4.2")
 
     testImplementation(project(":domain"))
+
 }
